Add GNOME stratum, system and cluster
This introduces a second version of libxkbcommon. This is unavoidable
really as (a) libinput requires libxkbcommon, (b) Mutter requires
libxkbcommon with X11 support enabled, and (c) X11 support in
libxkbcommon requires xcb-xkb. To avoid the input-common stratum
depending on the whole of X11, we do two variants of libxkbcommon and,
in the case where X11 support is needed, one overrides the other.

Change-Id: Ib6a47a4eb859d7b32c94f0cb05add774db6d15c8