Ensure newly brought up devices are correctly configured
This adds masking config for the virtual devices to prevent them from
attempting to DHCP, and stops us from giving the ip of eno1 to the
bridge device, since eno1's address was obtained by DHCP, so it's
against the rules to statically allocate that address to an interface.

Now, we DHCP for a new address for the bridge.