stage2-glibc: Fix symlink to dynamic linker
On non-x86 systems, the linker may not be called ld-linux.so.* but
the older name of ld.so.*

When the linker is ld.so.1, for example, the install script of
stage2-glibc does not exit with an error (as basename/ls are in
subshells), but rather $loader is empty string.

This means that the symlink made is from /lib/lib to /tools/lib
rather than from /lib/ld.so.1 to /tools/lib/ld.so.1.