Move patch link creation into openvswitch
This is required for Open vSwitch to be able to signal that every
network interface required has been configured.

It also means we no longer need to set the links to promiscuous mode
ourselves, since interfaces need to be set in promiscuous mode to allow
bridging to work and Open vSwitch handles this responsibility if it is
configured to be the one to do the link setup.