rawdisk.write: Add functions to read and process a partition specification
Read a partition specification, and create a dict containing the
required informtion to describe a partitioned device or image.

A file is used to specify the layout and properties of a partitioned
device or image. The specification is plain YAML, the path to which is
given as the value of PARTITION_FILE in a cluster definition.

There is an extant partitioning library for Python, pyparted, but this
relies on libparted, whereas the changes in this series implement a
simple wrapper for the fdisk command line tool, which is usually
available in Baserock.

Change-Id: Ia64e15e4c24e792e3244091340bed5febf5cea84