migrations/006: Report network errors more clearly
This is a bit of an awkward migration because it tries to contact the
remote cache server in order to do build system autodetection. Users who
are using a Trove other than git.baserock.org will need to change the
script to point to their Trove. So the error message should point them
in that direction, but without obscuring the original error (which may
due to network connectivity or soemthing else).

Error output now looks like this:

    WARNING: Unexpected response from server http://git.baserock.org:8080/
    for repo git://git.baserock.org/delta/libndp: 404 Client Error: Not
    Found
    Error: Unable to look up one or more repos on remote Git server
    git.baserock.org. If you are using a Trove that is not git.baserock.org,
    please edit the TROVE_HOST constant in this script and run it again.

Before, you would get a backtrace from the simplejson module in this
situation.

Change-Id: I3f6b3fb46ff436b8490a687ab4bd4d32d857f2ff