Scripts: Add system artifact manifest script
This script accepts an assembled Baserock system tarball, extracts metadata,
and resolves refs for the contained chunks, attempting to guess the version
for each. It then generates a report for the system.

Move this script, formerly a morph plugin, into definitions, and remove
dependency on morph.

Add some helper functions to replace morph functions, including an object
and methods which can be used to import metadata from an assembled system,
extensibly for conceivable metadata formats, including morph and ybd.

Change-Id: Iefc7911a339d450bfefdf3f8b13ee1a102ddb909