scripts/licensecheck.py: Fall back to `git clone` if morph isn't available
This is a quick fix to allow the script to be run without dependence on morph.
As discussed on IRC, since it is a more complex problem finding a way to allow
access to caches from different build tools, such a solution wasn't attempted.
When falling back to `git clone`, this change prevents the script from using
an existing cache, generating its own instead by cloning from the Basrock
cache server, and will take longer.

Also add automatic creation of REPOS_DIR if it doesn't exist.

Change-Id: Id4b0977a2cae90068b6c2e2fd0c62470287bdfcb